+++ R/spex.R
@@ -2,6 +2,7 @@
 #'
 #' Create Spatial Polygons with projection metadata from a 'Spatial Extent'.
 #'
+#' Called with no arguments will return the extent of the current 'par("usr")' setting. 
 #' This function is to replace a common pattern in the 'sp'-'raster' family which is
 #' \itemize{
 #' \item create an \code{\link[raster]{Extent}}, a bounding box in xmin,xmax,ymin,ymax but without projection metadata
@@ -51,8 +52,10 @@ spex <- function(x, ...) {
 
 #' @export
 #' @name spex
-spex.default <- function(x, byid = FALSE, .id, ...) {
-  if (byid) {
+spex.default <- function(x, crs, byid = FALSE, .id, ...) {
+  if (missing(crs)) crs <- NULL
+   if (missing(x)) return(spex(raster::extent(graphics::par("usr")), crs = crs))
+    if (byid) {
     stop("byid option not yet implemented")
     #lapply(split(x, seq(nrow(x))), raster::extent)
   } else {
@@ -70,3 +73,50 @@ spex.Extent <- function(x, crs, ...) {
   crs(p) <- crs
   spex(p, ...)
 }
+
+#' Extent of simple features
+#' 
+#' This is the simplest of the missing "raster support" for the sf package, 
+#' here using the xmin, xmax, ymin, ymax convention used by raster rather than 
+#' the transpose version favoured in sp and sf. 
+#' @param x object with an extent
+#' @param ... unused 
+#' @name extent
+#' @aliases Extent
+#' @importFrom raster extent
+extent_sf <- function(x, ...) {
+  raster::extent(attr(x[[attr(x, "sf_column")]], "bbox")[c(1, 3, 2, 4)])
+}
+setOldClass("sf")
+setMethod(f = "extent", signature = "sf", definition = extent_sf)
+
+#' @export
+#' @name spex
+spex.sf <- function(x, crs, ...) {
+  spex(extent(x), attr(x[[attr(x, "sf_column")]], "crs")$proj4string)
+}
+
+#' Axis ranges from extent
+#' 
+#' Functions `xlim` and `ylim` return the two-value counterparts of an extent. 
+#' 
+#' Any projection metadata is dropped since this is a one-dimensional entity. 
+#' @param x any object with an extent understood by `spex`
+#' @param ... reserved for future methods
+xlim <- function(x, ...) UseMethod("xlim")
+#' @export
+#' @name xlim
+xlim.default <- function(x, ...) {
+  spx <- spex(x)
+  c(raster::xmin(spx), raster::xmax(spx))
+}
+#' @export
+#' @name xlim
+ylim <- function(x, ...) UseMethod("ylim")
+#' @export
+#' @name xlim
+ylim.default <- function(x, ...) {
+  spx <- spex(x)
+  c(raster::ymin(spx), raster::ymax(spx))
+}
+